requirements.txt
keboola.component==1.3.7
keboola.utils
keboola.http-client
mock~=4.0.3
freezegun~=1.2
requests~=2.28.1